Shifted ContentTableModel and ContentTableCellRender classes
